backend/requirements.txt
# Core Dependencies
scikit-learn>=1.6.1
matplotlib>=3.10.0
pandas>=2.2.3
psycopg2-binary>=2.9.10
cryptography>=44.0.1
fastapi>=0.115.8
uvicorn>=0.34.0
pyautogen>=0.7.5
langfuse>=2.59.3
numpy>=1.19.5,<2.0.0
python-dotenv>=1.0.1
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
replicate>=0.22.0
together>=0.2.8
huggingface-hub>=0.21.3
fireworks-ai>=0.12.0
requests>=2.31.0
ray>=2.9.3
redis>=5.0.1

# AI Optimization
onnxruntime-gpu>=1.17.0  # For NVIDIA GPU support
redis>=5.0.1  # For caching

# Monitoring and Logging
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=7.0.0
loguru>=0.7.2
python-json-logger>=2.0.7
statsd>=4.0.1
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-prometheus>=1.23.0

# Testing Dependencies
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.5
pytest-mock>=3.12.0
httpx>=0.27.0
factory-boy>=3.3.0

# Code Quality
black>=24.1.1
flake8>=7.0.0
isort>=5.13.2
mypy>=1.8.0
pre-commit>=3.6.0
